requirements.txt
Pillow >= 5.1.0
requests >= 2.7.0
python-dateutil >= 2.6.1